Rewrite start-game
